dashboard/requirements.txt
streamlit>=1.42.0
plotly>=6.0.0
snowflake-connector-python>=3.12.0
pandas>=2.2.0
